backend/requirements.txt
openai
psycopg2-binary
python-dotenv
fastapi
uvicorn
pydantic
pdfplumber
python-docx
pandas
reportlab
sse-starlette
python-multipart
openpyxl
